得点計算を実装(閉路の内側も含む)

+++ src/useful.cpp
@@ -3,7 +3,7 @@
 UF::UF(uint_fast32_t size) {
 	data = std::vector<int_fast32_t>(size);
 	rank = std::vector<int_fast32_t>(size);
-	for(int_fast32_t i = 0; i < size; i++) {
+	for(size_t i = 0; i < size; i++) {
 		data[i] = i;
 		rank[i] = 0;
 	}
